+++ src/noahtheduke/splint/rules/style/prefer_var_dispatch.clj
@@ -4,18 +4,11 @@
 
 (ns ^:no-doc noahtheduke.splint.rules.style.prefer-var-dispatch
   (:require
+   [noahtheduke.splint.diagnostic :refer [->diagnostic]]
    [noahtheduke.splint.rules :refer [defrule]]
    [noahtheduke.splint.rules.helpers :refer [var??]]))
 
 (set! *warn-on-reflection* true)
-
-(defn not-var?
-  [obj]
-  (not (or
-         (string? obj)
-         (map? obj)
-         (and (list? obj)
-           (var?? (first obj))))))
 
 (defrule style/prefer-var-dispatch
   "During repl-driven development, one might wish to change the dispatch function of a multimethod. However, once set, the dispatch function cannot be changed without first unmapping the entire multimethod, which requires reloading all methods again.
@@ -31,6 +24,9 @@
   (def example-dispatch :type)
   (defmulti example #'example-dispatch)
   "
-  {:pattern '(defmulti ?name (?? ?docstring string?) (?? ?attr map?)
-               (? ?dispatch not-var?) (?* ?options))
-   :message "Use a var for the dispatch function to improve repl-drive development."})
+  {:pattern '(defmulti ?name ?*options)
+   :on-match (fn [ctx rule form {:syms [?name ?options]}]
+               (when-let [defmulti-form (:splint/defmulti-form (meta form))]
+                 (when-not (and (list? (:dispatch-fn defmulti-form))
+                             (var?? (first (:dispatch-fn defmulti-form))))
+                   (->diagnostic ctx rule form {:message "Use a var for the dispatch function to improve repl-drive development."}))))})